Split installer
